requirements.txt
streamlit
pandas
openpyxl
pypdf
python-docx
Pillow
pytesseract
